+++ www/IdmAuthFlows.js
@@ -233,6 +233,22 @@ var IdmAuthFlows = function() {
   };
 
   /**
+   * Utility method to change enumerability of keys.
+   */
+  var changeEnumberability = function(inputObject, nonEnumerableKeys)
+  {
+    var returnObject = {};
+    Object.keys(inputObject).forEach(function(keyName) {
+      isEnumerable = (nonEnumerableKeys.indexOf(keyName) > -1) ? false : true;
+      Object.defineProperty(returnObject, keyName, {
+        value: inputObject[keyName],
+        enumerable: isEnumerable
+      });
+    });
+    return returnObject;
+  };
+
+  /**
    * Utility method to validate number is greater than or equal to zero.
    */
   var assertPositiveOrZero = function(input, field)
@@ -1556,6 +1572,19 @@ var IdmAuthFlows = function() {
       return this;
     };
 
+    /**
+     * @function enableWebViewButtons
+     * @memberof OAuthPropertiesBuilder.prototype
+     * @param {Array.<string>} actionButtonList - List of buttons to be enabled in web view. Defaults to ['ALL'].
+     * @return {FedAuthPropertiesBuilder}
+     */
+    this.enableWebViewButtons = function(actionButtonList)
+    {
+      assertButtonsArray(actionButtonList, authPropertyKeys.EnableWebViewButtons);
+      this.put(authPropertyKeys.EnableWebViewButtons, actionButtonList);
+      return this;
+    };
+
     if (grantType)
       this.oAuthAuthorizationGrantType(grantType);
     if (tokenEndpoint)
@@ -2177,8 +2206,14 @@ var IdmAuthFlows = function() {
       }
       // End: Backwards compatibility
 
-      return new Promise(function (resolve, reject) {
+      var NON_ENUMERABLE_KEYS = ['ExpiryTime']; // ToDo: Pass this as a variable to backend
+
+      var getHeadersPromise = new Promise(function (resolve, reject) {
         exec(resolve, reject, TAG, 'getHeaders', [authFlowKey, opt.fedAuthSecuredUrl, opt.oauthScopes]);
+      });
+
+      return getHeadersPromise.then(function(response) {
+        return changeEnumberability(response, NON_ENUMERABLE_KEYS);
       });
     };
   };
@@ -2451,6 +2486,53 @@ var IdmAuthFlows = function() {
     var pinCallback = authProps[authPropertyKeys.PinChallengeCallback];
     var enablePromise, disablePromise;
     var self = this;
+
+    /**
+     * Store given [key,value] into local authenticator's secured keystore. There are two keystores available.
+     * First is the default keystore and second one is the local authenticator's keystore.
+     * Keys stored in default keystore can be retrieved even before the user is authenticated using local authenticator.
+     * This is suitable for storing app level preferences which are needed at app launch time before user login.
+     * Local authenticator keystore can be accessed only after successful local authentication.
+     * So user has to be logged in before data can be stored in this manner.
+     * Needless to say, second one is more secure than first. So apps should not store confidential information using first method.
+     *
+     * Using same key across default and local authenticator keystores are not allowed.
+     * The value last set will be the one that will be stored and the value set into other keystore will be lost.
+     * For e.g., if you set ("foo","bar") in default authenticator and then after login, set ("foo","hello"), then when 
+     * you try to get back the key "foo" you will get the value as "hello".
+     * @function setPreference
+     * @memberof LocalAuthenticationFlowManager.prototype
+     * @param {String} key - key to store
+     * @param {String} value - value to store
+     * @param {boolean} secure - false to store in default keystore. Defaults to true.
+     * @return {Promise.<LocalAuthPropertiesBuilder.LocalAuthenticatorType>}
+     * If the promise is rejected, the callback will receive and object of type {@link AuthError}
+     */
+    this.setPreference = function(key, value, secure) {
+      assertString(key, "key");
+      return new Promise(function(resolve, reject) {
+        exec(resolve, reject, TAG, 'setPreference', [id, key, value, secure])
+      });
+    };
+
+    /**
+     * Fetches given key's value from local authenticator's keystore.
+     * Firstly user authentication is checked. If user is autehnticated using local authenticator, a check is performed
+     * on secured keystore. If key is found there its corresponding value is being returned. If key is not found in secured
+     * keystore, it fallsback to default keystore and searches the key overthere.
+     * If user is not authenticated, key will be searched from default keystore.
+     * @function enable
+     * @memberof LocalAuthenticationFlowManager.prototype
+     * @param {String} key - key to fetch
+     * @return {Promise.<LocalAuthPropertiesBuilder.LocalAuthenticatorType>}
+     * If the promise is rejected, the callback will receive and object of type {@link AuthError}
+     */
+    this.getPreference = function(key) {
+      assertString(key, "key");
+      return new Promise(function(resolve, reject) {
+        exec(resolve, reject, TAG, 'getPreference', [id, key])
+      });
+    };
 
     /**
      * Get all enabled local authenticator types, in primary first order.